Removed unnecessary variables, reduced repetition
The variables contained in the DoLE object were there as remnants to prior bookmarklet code. They're not needed as we're using a full JS file with functions. We've also gone through and replaced the full object in DoLCleanBody() with a for loop. This plays double duty, in reducing code repetition, and preventing any issues if other liquids are added to the game in future. We don't want to be accidentally deleting vanilla game variables.